+++ terraform/modules/happy-service-ecs/main.tf
@@ -25,6 +25,111 @@ resource "aws_ecs_service" "service" {
 locals {
   task_definition = [
     {
+      name      = "datadog-agent"
+      essential = true
+      image     = "${var.datadog_agent.registry}:${var.datadog_agent.tag}"
+      cpu       = var.datadog_agent.cpu
+      memory    = var.datadog_agent.memory
+
+      environment = concat(
+        [
+          {
+            name  = "DD_API_KEY"
+            value = var.datadog_api_key
+          },
+          {
+            name  = "DD_SITE"
+            value = "datadoghq.com"
+          },
+          {
+            name  = "DD_SERVICE"
+            value = var.app_name
+          },
+          {
+            name  = "DD_ENV"
+            value = var.deployment_stage
+          },
+          {
+            name  = "ECS_FARGATE"
+            value = "true"
+          },
+          {
+            name  = "DD_APM_ENABLED"
+            value = "false"
+          },
+          {
+            name  = "DD_DOGSTATSD_NON_LOCAL_TRAFFIC"
+            value = "true"
+          },
+          {
+            name  = "DD_APM_NON_LOCAL_TRAFFIC"
+            value = "true"
+          },
+          {
+            name  = "DD_PROCESS_AGENT_ENABLED"
+            value = "true"
+          },
+          {
+            name  = "DD_RUNTIME_METRICS_ENABLED"
+            value = "true"
+          },
+          {
+            name  = "DD_SYSTEM_PROBE_ENABLED"
+            value = "false"
+          },
+          {
+            name  = "DD_GEVENT_PATCH_ALL"
+            value = "true"
+          },
+          {
+            name  = "DD_APM_FILTER_TAGS_REJECT"
+            value = "http.useragent:ELB-HealthChecker/2.0"
+          },
+          {
+            name  = "DD_TRACE_DEBUG"
+            value = "true"
+          },
+          {
+            name  = "DD_LOG_LEVEL"
+            value = "debug"
+          },
+          {
+            name  = "DD_EXPVAR_PORT"
+            value = "6000"
+          },
+          {
+            name  = "DD_CMD_PORT"
+            value = "6001"
+          },
+          {
+            name  = "DD_GUI_PORT"
+            value = "6002"
+          }
+      ])
+
+      "port_mappings" = [
+        {
+          containerPort = 8126
+          hostPort      = 8126
+          protocol      = "tcp"
+        },
+        {
+          containerPort = 8125
+          hostPort      = 8125
+          protocol      = "udp"
+      }]
+
+      "logConfiguration" = {
+        logDriver = "awslogs"
+
+        options = {
+          awslogs-stream-prefix = var.app_name,
+          awslogs-group         = aws_cloudwatch_log_group.cloud_watch_datadog_agent_logs_group.id,
+          awslogs-region        = data.aws_region.current.name
+        }
+      }
+    },
+    {
       name              = var.app_name
       essential         = true
       image             = var.image
@@ -67,6 +172,16 @@ locals {
           awslogs-region        = data.aws_region.current.name
         }
       }
+      dockerLabels = {
+        # TODO: Add happy_owner and happy_project tags
+        "com.datadoghq.ad.tags" = jsonencode([
+          "happy_stack:${var.tags.happy_stack_name}",
+          "happy_service:${var.tags.happy_service_name}",
+          "deployment_stage:${var.deployment_stage}",
+          "env:${var.tags.happy_env}",
+          "service:${var.tags.happy_service_name}"
+        ])
+      },
     }
   ]
 }
@@ -86,6 +201,12 @@ resource "aws_ecs_task_definition" "task_definition" {
 resource "aws_cloudwatch_log_group" "cloud_watch_logs_group" {
   retention_in_days = 365
   name              = "/${var.stack_resource_prefix}/${var.deployment_stage}/${var.custom_stack_name}/${var.app_name}"
+  tags              = var.tags
+}
+
+resource "aws_cloudwatch_log_group" "cloud_watch_datadog_agent_logs_group" {
+  retention_in_days = 365
+  name              = "/${var.stack_resource_prefix}/${var.deployment_stage}/${var.custom_stack_name}/${var.app_name}/datadog-agent"
   tags              = var.tags
 }