+++ terraform/modules/happy-service-ecs/variables.tf
@@ -142,3 +142,25 @@ variable "tags" {
   })
   description = "The happy conventional tags."
 }
+
+variable "datadog_api_key" {
+  type        = string
+  default     = ""
+  description = "DataDog API Key"
+}
+
+variable "datadog_agent" {
+  type = object({
+    registry : optional(string, "public.ecr.aws/datadog/agent"),
+    tag : optional(string, "latest"),
+    memory : optional(number, 512),
+    cpu : optional(number, 512)
+  })
+  default = {
+    registry = "public.ecr.aws/datadog/agent"
+    tag      = "latest"
+    memory   = 512
+    cpu      = 512
+  }
+  description = "DataDog agent image to use"
+}