+++ terraform/modules/happy-stack-ecs/main.tf
@@ -16,23 +16,23 @@ locals {
   vpc_id             = local.secret["vpc_id"]
   subnets            = local.secret["private_subnets"]
   security_groups    = local.secret["security_groups"]
-  zone               = local.secret["zone_id"]
   cluster            = local.secret["cluster_arn"]
   ecs_execution_role = lookup(local.secret, "ecs_execution_role", "")
+  datadog_api_key    = try(local.secret["datadog_api_key"], "")
 
   image = join(":", [local.secret["ecrs"][local.app_name]["url"], lookup(var.image_tags, local.app_name, var.image_tag)])
 
   external_dns = local.secret["external_zone_name"]
   internal_dns = local.secret["internal_zone_name"]
-  dns_zone_id  = local.secret["zone_id"]
 
   listener_arn = local.secret[local.alb_key][local.app_name]["listener_arn"]
   alb_zone     = local.secret[local.alb_key][local.app_name]["zone_id"]
   alb_dns      = local.secret[local.alb_key][local.app_name]["dns_name"]
 
+
+
   ecs_role_arn  = local.secret["service_roles"]["ecs_role"]
   ecs_role_name = element(split("/", local.secret["service_roles"]["ecs_role"]), length(split("/", local.secret["service_roles"]["ecs_role"])) - 1)
-  url           = try(join("", ["https://", module.dns[0].dns_prefix, ".", local.external_dns]), var.url)
 
   stack_resource_prefix = local.app_name
 
@@ -85,4 +85,5 @@ module "service" {
   additional_env_vars   = local.db_env_vars
   chamber_service       = var.chamber_service
   tags                  = local.tags
+  datadog_api_key       = local.datadog_api_key
 }