+++ terraform/modules/happy-tfe-okta-app/outputs.tf
@@ -1,0 +1,3 @@
+output "oidc_config" {
+  value = module.happy_app.oidc_config
+}