+++ terraform/modules/happy-env-eks/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [2.0.0](https://github.com/chanzuckerberg/happy/compare/happy-env-eks-v1.6.0...happy-env-eks-v2.0.0) (2023-01-27)
+
+
+### ⚠ BREAKING CHANGES
+
+* add oidc configuration to happy env,int secret ([#1020](https://github.com/chanzuckerberg/happy/issues/1020))
+
+### Features
+
+* add oidc configuration to happy env,int secret ([#1020](https://github.com/chanzuckerberg/happy/issues/1020)) ([d887dff](https://github.com/chanzuckerberg/happy/commit/d887dff7755a6899e2cf09e592a70b906ae53671))
+
+## [1.6.0](https://github.com/chanzuckerberg/happy/compare/happy-env-eks-v1.5.0...happy-env-eks-v1.6.0) (2023-01-24)
+
+
+### Features
+
+* (CCIE-1004) Enable creation of stack-level ingress resources with a context based routing support ([#986](https://github.com/chanzuckerberg/happy/issues/986)) ([f258387](https://github.com/chanzuckerberg/happy/commit/f258387b72c1a0753c2779a79b0de8da56df71f1))
+
+## [1.5.0](https://github.com/chanzuckerberg/happy/compare/happy-env-eks-v1.4.0...happy-env-eks-v1.5.0) (2023-01-24)
+
+
+### Features
+
+* add tags from integration secret ([#990](https://github.com/chanzuckerberg/happy/issues/990)) ([46fcd8a](https://github.com/chanzuckerberg/happy/commit/46fcd8a99118b70add0feaecc0d9dd4358100bf0))
+
 ## [1.4.0](https://github.com/chanzuckerberg/happy/compare/happy-env-eks-v1.3.0...happy-env-eks-v1.4.0) (2023-01-09)